+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `draw_polylines` to `compas_rhino.artists.Artist`.
+- Added `color` argument to `compas_rhino.artists.MeshArtist.draw_mesh`
+
 ### Changed
 
 - Fix setting of environment variables in `System.Diagnostics.Process`-based subprocess for `XFunc` and `RPC`.